+++ src/rtsp/src/rl.cpp
@@ -106,7 +106,7 @@ av_cold void ff_rl_init_vlc(RLTable *rl, unsigned static_size)
 {
     int i, q;
     VLC_TYPE table[1500][2] = {{0}};
-    VLC vlc = {  .bits = 0, .table = table, .table_size = 0,  .table_allocated = static_size };
+    VLC vlc = {  .bits = 0, .table = table, .table_size = 0,  .table_allocated = (int)static_size };
     av_assert0(static_size <= FF_ARRAY_ELEMS(table));
     init_vlc(&vlc, 9, rl->n + 1, &rl->table_vlc[0][1], 4, 2, &rl->table_vlc[0][0], 4, 2, INIT_VLC_USE_NEW_STATIC);
